Restarted some stuff so I can do it better. Made a reader that can reade across data blocks if necessary Still can't get things to read right

+++ internal/inode/util.go
@@ -3,8 +3,8 @@ package inode
 //ProcessInodeRef processes an inode reference and returns two values
 //The first value is the inode table offset. AKA, it's where the metadata block of the inode STARTS.
 //The second value is the offset of the inode, INSIDE of the metadata.
-func ProcessInodeRef(inodeRef uint64) (tableOffset uint32, metaOffset uint16) {
-	tableOffset = uint32(inodeRef >> 16)
-	metaOffset = uint16(inodeRef &^ 0xFFFFFFFF0000)
+func ProcessInodeRef(inodeRef uint64) (tableOffset uint64, metaOffset uint64) {
+	tableOffset = inodeRef >> 16
+	metaOffset = inodeRef &^ 0xFFFFFFFF0000
 	return
 }
